refactor(backend): group blob cleanup workflow
