+++ apps/backend/src/entrypoints/scheduledJobs/lambda-generated-media-promotion.ts
@@ -7,7 +7,7 @@ import {
   MediaBlobCleanupBatchError,
   runMediaBlobCleanupBatch,
   type MediaBlobCleanupBatchResult,
-} from "../../mediaAssets/blobLifecycle/cleanupProcessor";
+} from "../../mediaAssets/blobLifecycle/cleanup/processor";
 import {
   addBackendBreadcrumb, captureBackendException, createBackendObservationScope,
   type BackendObservationScope,
