+++ apps/backend/src/mediaAssets/blobLifecycle/cleanup/repository.ts
@@ -1,6 +1,6 @@
-import type { DatabaseExecutor } from "../../database";
-import { unsafeTransactionWithDeadline } from "../../database/unsafe";
-import { buildMediaBlobStorageKey } from "../storageKeys";
+import type { DatabaseExecutor } from "../../../database";
+import { unsafeTransactionWithDeadline } from "../../../database/unsafe";
+import { buildMediaBlobStorageKey } from "../../storageKeys";
 
 const sha256Pattern = /^[0-9a-f]{64}$/u;
 const uuidPattern =
